Eliminate (semantically) a repeated search in RemoveViewProp
Replaced HasViewProp() call with IndexOfFirstOccurence(), the former is just a thin wrapper of the latter anyway.

This way we can call the other variant of RemoveItem(), the one that takes the index instead of the prop itself.

With the current implementation of RemoveItem, this doesn't help much, but eventually can, as now it doesn't have to search for the prop *again*.