Replaced the only call to RemoveActor2D with RemoveViewProp
This is a slight behaviour change because the former not only calls the latter, but also calls `Actors2D->RemoveItem`.

However, the `Actors2D` list seems to be only accessible from `vtkViewport::GetActors2D` and the first thing it does is empty `Actors2D` and rebuild it.

So I think this change is safe.