Make vtkCameraOrientationWidget robust for deserialization
- This commit is a first step to getting the widget to deserialize
correctly because during deserialization a different renderer
is assigned as the default renderer.
- When another renderer is provided, this commit carefully removes
the previous reorientation observer. If the widget was previously
enabled, it replaces the older renderer in the render window with the
new renderer. Otherwise, upon deserialization, the render window had
an extra renderer which originated from the constructor of
`vtkCameraOrientationWidget``.
- Updated baseline images which used the camera orientation widget
  because the widget now resizes itself to correct proportions