vtkSocket: fix win32 error handling
Previously when encountering an error during a Receive() call, server
hangs indefinitely as only WSAENOBUFS was checked.

For WSAECONNABORTED we cannot know if the disconnection is expected or
not so we don't output an error but log a trace for debugging purpose.

(cherry picked from commit c86e3aaaf9b3f2125764b40cc8b40b80b8700b99)