Warn when default __init__() modifies C++ attributes
If the C++ object already existed within VTK, and the Python object
is being created because it is being returned by a method, then we
don't want the creation of the Python object to cause changes to
C++ object attributes that were already set by the C++ VTK code.

For example, if we are getting the output dataset from a C++ VTK
filter, we don't want any C++ attributes of the dataset to change
when the Python part of this pre-existing dataset object is created.