Introduce vtkExplodeDataSet: create composite out of cell data id
vtkExplodeDataSet creates a vtkPartitionedDataSetCollection from any input dataset according to a cell scalar isovalue criteria.
Each partition of the output contains cells that share the same value for the given cell array.

It comes in replacement of vtkSplitByCellScalarFilter to use
vtkPartitionedDataSetCollection instead of vtkMultiBlockDataSet as output.
Also it relies on the threaded vtkExtractCells filter.