vtkImageSSIM: handle SSIM's [-1,1] range
When SSIM returns a negative value, the histogram bucketing ends up
writing to out-of-bounds memory. Don't do that and instead shift the
buckets to account for the full range of values that are possible.

Finally, add commentary on the code.

The clamping is still preserved as enabled by default in vtkTesting,
because we want to treat the anti-correlated images (negative values)
in the same way as not similar.