vtkTesting: handle SSIM's [-1,1] range
When SSIM returns a negative value, the histogram bucketing ends up
writing to out-of-bounds memory. Don't do that and instead shift the
buckets to account for the full range of values that are possible.

Finally, add commentary on the code.

See commit cd75f4b9618cb62e623c4adf88c5da0b346e2a83 which applied this
to `master` and `release`.

Co-authored-by: Spiros Tsalikis <spiros.tsalikis@kitware.com>