AMReXParticlesReader: handle inexact division of grids among MPI ranks
- this commit fixes the grid-process assignment logic robust
when the number of grids cannot be exactly divided among
the number of MPI ranks (pieces in this code)
- this commit also adds MPI awareness to an existing unit test.
- the unit test TestAMReXParticlesReader is run with MPI, when enabled,
on 3 nodes. It passes only when the sum of the number of points across
all nodes is equal to a known value.
- without the change in vtkAMReXParticlesReader, the unit test will faill with MPI

(cherry picked from commit de94dcedf7b4a108caf7b0c552bbf221da242ac8)