Fix issue with Fides initializing HIP before using
There was an issue with Fides crashing because it attempted to allocate
memory with HIP before HIP was initialized. (See
https://gitlab.kitware.com/vtk/fides/-/issues/26.) The problem was
actually with VTK-m. (Initializing VTK-m is supposed to be optional,
but HIP was not getting initialized in time if it was not.) To fix
the Fides issue, update VTK-m to the changes provided by
https://gitlab.kitware.com/vtk/vtk-m/-/merge_requests/3289.