vtkModuleWrapPython: support depending on modules being wrapped
The `TARGET` argument is an `INTERFACE` library which does not end up in
the build graph as something that can be requested of `make` or `ninja`
or works as a dependency of a custom target. Instead, make a custom
target that can be requested at build time by the user and depended upon
by other custom targets in the build.