Restore vtkGeometryFilter behavior for degenerated cells
`Initialize` is expected to put lowest point id first.
This was not the case with a face that have duplicated point.

This is a regression starting from commit 4a46c5dd (MR !10025)

Try with the degenerated tri `1 0 0`.
As `0` is not strictly lower than the second `0`,
the algo keeps `1` (first id of the list) as the lowest.
Using `<=` correct this behavior.

Without this change, vtkGeometryFilter misses some external faces.
See #19600 for more discussion.