Fix light transforms due shadow map computation
The vtkShadowMapBakerPass computes the shadow map by simulating a camera
for each light in the scene and compositing the rendered result. The
simulation inadvertently changes the original light parameters. This
change caches the original light transform and resets it after the
shadow computation to ensure that successive renders using the same
light are correct.