java: refer java1.8 as java8 in download_java.com
This commit uses a fixed JDK version for every JAVA build which is
independent of the value of `VTK_JAVA_VERSION`, this possible since a
newer JDK can produce an older JAVA bytecode using the -release flag
(that we use).

Now `VTK_JAVA_VERSION`only decides the JAVA version of the produced JAVA
bytecode.