vtkGenericEnSightReader: rewrite debug assertion avoidance predicate
This predicate clause was introduced in e6d1e71edc (BUG: Avoiding debug
assertion on with VS8…, 2009-05-06) in the context of avoiding a debug
assertion in VS8 when using the routine during `CanReadFile` as it opens
in ASCII mode but might actually be a binary file.

On aarch64, this way of testing the high bit is a tautology. Instead,
use the `isascii` function to make the expression meaningful at a
glance.