vtkMNITagPointReader: fix integer parsing
This code was rewritten in dda651582f (BUG: Fix escape character
handling, improve test coverage., 2010-11-24). Given the `- '0'` on the
value guarded by the conditionals, it is likely that these were meant to
be the ASCII value of the numerals 0-9, not the numeral values 0-9. Use
`isdigit` to simplify the expression.

Discovered via Linux aarch64 warnings about `*pos >= 0` being a
tautology.