Deprecate interactor WindowResizeEvent
The vtkRenderWindowInteractor generates ConfigureEvent when it
is notified by the UI that the window size has changed (the
ConfigureEvent was added in VTK 3.2), so the recent addition
of WindowResizeEvent to the interactor was unnecessary and has
caused problems because it was not correctly generated by X11.

The ConfigureEvent, unlike WindowResizeEvent, is observed by
and customizable via the vtkInteractorStyle classes.

The ConfigureEvent itself is not deprecated, only its use with
vtkRenderWindowInteractor is deprecated.  This event was added
in VTK 8.2 to check when vtkWindow::Size is modified, and it
is still serves this purpose (particularly for programs that do
not have an interactor).