vtkIOFides: Declare module test dependencies
Adding VTK::FiltersParallelMPI dependency to VTK::FiltersParallelDIY2
caused a macos kits build in CI to skip building VTK::FiltersParallelDIY2,
which in turn caused the VTK::IOFides basic writer test to fail when
it could not import the module.

The VTK::IOFides module is missing a declaration of its test dependency
on VTK::FiltersParallelDIY2.  Declaring that dependency seems to solve
the problem described above.