PIO: load dump files in a ordered manner
This is needed since in some machines (VALL) the order in which directories are
listed put hash-stamp files first. This is particularly worrisome in linux
machines since readdir() does not guarantee the order of the listed files. This
results in the PIO reader to fail to guess the type of the file (PIO or HDF5).

Error observed in ParaView:
https://gitlab.kitware.com/paraview/paraview/-/jobs/11228146