Remove VTKJavaWrapped file from vtkParseJava output
Commit 048e7d992be, added in VTK v4, caused vtkParseJava to write
a file called VTKJavaWrapped to the output directory.  This file
was, essentially, used as a timestamp for the CMake rule for
building .java files for the wrappers.  The current cmake rules
for the wrappers no longer use this file.  In modern builds where
vtkParseJava is run in parallel, having each vtkParseJava process
attempting to write to this file is unwise.