vtkProbeLineFilter: Use builtin locators when possible
When handling vtkPointSet derived classes use the cell-locator of the dataset
instead of an external one.  This avoids rebuilding cell locators and
simplifies lifetimes and ownership of locators. Untangling these lifetimes
fixes memory leaks in the previous implementation that could occur with
unstructured grid inputs when using the SAMPLE_LINE_UNIFORMLY sampling pattern.