vtkHDFReader: fix composite ImageData reading
ImageData had not been adapted for composite reading. Extent and spacing were read only once during RequestInformation, and stored to be used afterwards. This does not work in a composite setting, where the top-level block is not an ImageData. Now, we read those properties again when reading the block itself, to make sure they are up to date, and do not store them at the class level anymore.