Fix extent for vtkSynchronizedTemplates2D
This adjusts the code for setting the execution extent to be the
same as vtkSynchronizedTemplates3D.  Now, the execution extent
is the intersection of the Extent and the UpdateExtent.

The purpose is to give correct behavior when the WholeExtent is
smaller than the Extent, which is common after image clipping.
The WholeExtent constrains the UpdateExtent to within the clipped
region, and we want to respect that when we execute the filter.
This is, in fact, crucial if the Extent is 3D (a stack/volume)
but the clipped WholeExtent is 2D (a slice) since 2D contouring
only works on slices.

Also, use vtkIdType to store the product of dimensions, to allow
contouring of large images.