maven: parametrize env name for upload deployments
Currently there is only a single enviroment for each Java version. This
results in release and master branch sharing the same deployment env
which in turn results in uploads jobs in release to sometimes fail to
deploy as they are considered outdated since a more recent child commit
in master has already been deployed.

backport: release