Fix some incorrectly wrapped GetNumberOf methods
The GetNumberOf methods, like GetNumberOfPoints, should only be
marked as a GET_NUMBER_OF method by vtkParseProperties.c if there
is a matching indexed Get method, like GetPoints(i).

The logic for this was present in vtkParseProperty.c, but it had
a bug.  As a result, the vtkExtentTranslator::GetNumberOfPieces()
method was assigned to the wrong property: its correct property
is NumberOfPieces rather than simply Pieces, because the
vtkExtentTranslator::GetPiece() method does not take an index
as an argument.

There was an additional bug in vtkParseProperties, in that it
was checking for VTK_PARSE_ID_TYPE which is no longer used by
the wrappers.  Instead, it needs to check the types that vtkIdType
expands to: VTK_PARSE_LONG_LONG and VTK_PARSE_INT.