Python: add a new struct for wrappable functions
Previously, the Python wrappers would modify the FunctionInfo data
in the following way while wrapping:
1. setting "Name" to NULL for functions that wouldn't be wrapped
2. converting C++ method signatures to Pythonic form for docstrings

To avoid modifying the FunctionInfo, the WrappedFunction struct
has been added.  Each WrappedFunction stores a pointer to the
unmodified FunctionInfo, along with the Pythonic form of the method
signature.

This change has no impact on the wrapper code that is generated.