Python: C code styling of ConvertPascalToSnake
Changed comments to C style for consistency with rest of file.
Also removed the NULL checks.  If property name is NULL, the
property struct has been corrupted and it's better to segfault
than to skip the property.  Likewise, if we run out of memory
during the build its better to segfault than to skip properties.