Fix wrap property ambiguity for SetValue(int, int)
When generating properties for wrapping SetValue(int, int)
or SetValue(vtkIdType, vtkIdType), there is ambiguity about
whether the first parameter is meant to be an index or not.
This patch resolves the ambiguity by examining the matched
GetValue() method.  For example:

    void SetId(vtkIdType i, vtkIdType id)
    vtkIdType GetId(vtkIdType i) // indexed

    void SetCorner(int x, int y)
    int* GetCorner() // not indexed

If the Get method is indexed, then the Set method is likewise
assumed to be indexed if its first argument is an integer type.