Initial pass of clang-tidy modernize-use-constexpr, trivial changes only
This changes some `const` to `constexpr` for trivial scalars or strings. No changes to functions.

In some cases it repositioned the `constexpr` to before the type, instead of after, which seems the more common case in the codebase anyway.

All these changes were made automatically by clang-tidy, but I chose this subset of its changes.