RenderingParallel: Support multiple simultaneous CAVE viewers
Add a class to manage a collection of independent viewers, each of
which is represented by an eye transform matrix and an eye separation
value.

Update vtkSynchronizedRenderers to send/receive this collection of
independent viewers, if present and not empty, during rendering.

Update TestClientServerRendering to validate that independent
viewer data are properly transferred from client to server.