Refactor statistics algorithms: output a custom data type…
… for statistical models. A `vtkStatisticalModel` is a collection
of tables organized by the "phase" of a statistics-algorithm
which created them.

This commit adds methods which enable any algorithm to
own an instance of `vtkStatisticsAlgorithmPrivate` (which
holds sets of requested attribute-array names) and copy its
contents to a `vtkStatisticsAlgorithm` subclass they own.
This way, a "wrapper" algorithm that adapts a `vtkDataObject`
into a `vtkTable` can run on *any* statistics algorithm
(instead of requiring a subclass of the wrapper for each
subclass of vtkStatisticsAlgorithm as was done in the past).

We also add `vtkModelData` which adapts geometric and composite
VTK data into tables and creates statistical models of them.
It accepts composite data held in a distributed-memory environment
and performs communication as needed to build one model (or one
hierarchy of models for composite data if requested) across all
ranks. It does not currently handle multiblock datasets (you should
convert to a partitioned dataset collection instead) or cell grids
(though there is some work in this direction).

Because `vtkModelData` owns and can apply an instance of any
subclass of `vtkStatisticsAlgorithm` to the tables it produces, the
`vtkStatisticsAlgorithm` class now implements methods to
serialize and deserialize ivars of any subclass. This allows
a `vtkStatisticalModel` (which stores the serialization) to
produce an instance of a properly-configured `vtkStatisticsAlgorithm`
to perform further processing (such as assessing data or testing the
likelihood of a model).

In order to facilitate communication of statistical model data
across ranks, both an XML and a legacy reader/writer pair for
the new vtkStatisticalModel data object are added.
The XML reader/writer only perform inline ASCII reads/writes for now.
Models are small, so that is all that should be required.

We add a new statistics algorithm for combined order+moment stats
called `vtkVisualStatistics`. It inherits `vtkDescriptiveStatistics`
and adds a fixed-width bin approximation to a histogram of data.
This has the potential to be faster than `vtkOrderStatistics` but
cannot capture unique frequent values the way order statistics can.

A new filter, `vtkSumTables` is provided to sum numeric columns
of a set of input `vtkTable` instances. This is used by
`vtkVisualStatistics` to aggregate binned histogram models.

`vtkExtractStatisticalModelTables` is a new filter to extract
model tables from a statistical model so that applications which
do not deal with `vtkStatisticalModel` data objects can still
access model information.

We centralize and expand ghost counting for statistics:
+ Move GhostsToSkip bit-mask and NumberOfGhosts ivar to the
  base `vtkStatisticsAlgorithm` class. This removes a lot of
  duplication in subclasses. Really, all subclasses should
  support skipping ghosts.
+ Expand support for skipping ghosted samples to the
  `vtkCorrelativeStatistics` filter.