Update the vtkParse generation instructions
Add the NOLINTNEXTLINE comment to the instructions, and remove the
instructions from vtkParse.y itself since it's best to have just
the one copy of the instructions in the README file.

The file vtkParse.tab.c has been regenerated to account for the
changed line numbers of the rules in vtkParse.y.