Refactor attribute handling out of vtkParse.y
This makes it possible to add new custom attributes without having
to regenerate the parser.  The list of attributes and the functions
for handling them are now in vtkParseAttributes.c.

Attributes with different roles (declaration vs. class vs. id) are
properly separated now, so the use of various attributes within a
single expression won't cause the attributes to be incorrectly
assigned.