Revert "Allow initializing addons in object manager python"
This reverts commit caae346a098ad288fa35cd1b0e1c1c82c2ab35e3.

This commit is reverted because it broke building python wrappers
when VTK_WRAP_SERIALIZATION is OFF. It can be fixed by tweaking the
generator expression in vtkModuleWrapPython.cmake to consider
VTK_WRAP_SERIALIZATION. But we will be pushing a new logic shortly
for serdes initialization that is better than current logic.
It will not require the python wrappers to know about
serdes.