Use lambertian diffuse factor for image based illumination
In the Cook-Torrance BRDF, the diffuse illumination component uses a
lambert factor. This factor is simply the albedo color divided by pi.

The shader currently uses this lambertian diffuse factor for direct
illumination but was missing for image based illumination.

So, we're putting the diffuse lambert term into the image-based lighting
diffuse component. This will have an affect of cutting albedo to 1/3-ish
(which is not the same as cutting full illumination to 1/3 of the
original, because there is more that affects the final image than just
this lambertian factor). Image-based diffuse effects will be dimmer.

References:
  https://learnopengl.com/PBR/Theory
  https://google.github.io/filament/Filament.html#materialsystem/diffusebrdf
  https://learnopengl.com/PBR/IBL/Diffuse-irradiance (the term c/π is the lambertian diffuse component)
  https://google.github.io/filament/Filament.html#toc5.3.4.6 (shown in the implementation as Fd_lambert().)