vtkIOSSReaderInternal: Consider 0 extents for StructuredBlock
I believe what we found is that when a block had no data it was getting
this point with extents of 0,0,0,0,0,0, which actually indicated that it
had one item, and so the following code core dumped because it tried to
copy an element or node when in fact there was none. The fix here is to
detect when we have 0,0,0,0,0,0 extents and switch it to 0,-1,0,-1,0,-1
to indicate truly no data. There may be a better way to do this, perhaps
by initializing the extents differently than how they are initialized now,
but this change definitely fixes all crashes in all my testing, and it
does not break any tests that I know about.