RedistributeDataSet : fix missing cells in ASSIGN_TO_ONE_REGION mode
Cells that lie on the boundary of multiple partitions could be deleted
when calling the extract cells filter. More precisely it will occur only
when this is a concave cell. Previously this use case was supported but
has been removed during a refacto of the details::GenerateCellRegions()
method. This fix reintroduce this case.

This is a regression between vtk 9.2.0 and vtk 9.3.0 during the refactor
of the redistribute dataset filter during this MR:

https://gitlab.kitware.com/vtk/vtk/-/merge_requests/9807
(cherry picked from commit 805191a5f993553a571ae770f755f0f0b823c3d0)