Fix gap when streaming slices with vtkImageReslice
When vtkImageReslice was used in a streaming pipeline, that is,
a pipeline created with SetInputConnection() both upstream and
downstream of vtkImageReslice, there could be a gap with a width
of 0.000008 times the slice spacing between adjacent slices
that were streamed through the pipeline.  For example, this gap
would show as a black (zero) output slice when the reslice
transform translated the volume by 0.499999 slice spacings.
This patch adds a small tolerance to the interpolator bounds
check in order to eliminate the gap.

This bug has been present since VTK 6, when the bounds checking
switched from being based on the WholeExtent of the volume, to
being based instead on the Extent of the data being streamed.