Fix image interpolator bug for update extents
The code that returned the image interpolator kernel size,
specifically for use in computing how much larger the upstream
extent needed to be than the downstream extent, had a bug where
sometimes the upstream extent was not padded in all three
directions.  This was due to an incorrect optimization caused
by a range error on a loop.  The lack of sufficient input
slices for interpolation could cause the output interpolation
in e.g. the slice direction to be of lower order than desired.