Fix issue in vtkStaticPointLocator in IntersectWithLine function
We want to store the closest point we hit in tMin. The objective of this condition was to check if we hit a point that is closest to the one we previously hit.
The distance to line was calculated into tHit but we were comparing tMin to t instead of tHit. So I fixed the comparison and the update of tMin with the value of tHit we just calculated.