Account for VTK_WHEEL_BUILD for RenderingTk suffix
The vtkLoadPythonTkWidgets module was not checking VTK_WHEEL_BUILD,
which causes the version suffix to be removed from all libraries.
As a results, the tk widget could not be used from VTK wheels,
unless VTK_VERSIONED_INSTALL was explicitly turned off.

Also, the cmake variable checks in vtkLoadPythonTkWidgets have been
changed to case-insensitive, to match the way cmake treats booleans.