Fix inappropriate reference to null array
A recent change to how cells are referenced in unstructured grids
(!12452/f4a0bf7b0969dede74f53c40b920c821f3581c11) introduced an error in
`vtkExtractCellsAlongPolyLine` where code that used a nullptr to signal
the use of polydata instead of unstructured data caused the array to be
referenced. This fixes the optimization to remove invalid references to
that array.