vtkOpenGLGL2PSHelperImpl: Include GL
`vtkOpenGLGL2PSHelperImpl` contains a direct call to `glGetDoublev`, but
was not including any headers for the same. This caused it to rely on
gl2ps implicitly including GL's headers, and also resulted in a link
error because the GL library is not linked. Fix this by including
`vtk_glad.h`, which allows the function to use the same delayed-linking
used by the rest of VTK.