vtkIOSSReader in Catalyst mode
Added set/get method for the Catalyst Conduit channel name.

Reader will now try and get the first node block if a node
block called "nodeblock_1" is not found. Will throw a
runtime exception if the region node block list is empty
when reading unstructured data.

Added a parallel test for structured and unstructured data on four
MPI ranks with one cell in the mesh. The test changes the channel name
from the default value and uses a node block name that is not
"nodeblock_1" for the unstructured case. Test also verifies that
structured data is handled properly when a rank has no data.